fix_typeError_in_dataloaders

+++ utils/dataloaders.py
@@ -724,7 +724,8 @@ class LoadImagesAndLabels(Dataset):
 
         labels_out = flow.zeros((nl, 6))
         if nl:
-            labels_out[:, 1:] = flow.from_numpy(labels)
+            # labels_out[:, 1:] = flow.from_numpy(labels)
+            labels_out[:, 1:] = flow.from_numpy(labels).to(labels_out.dtype)
 
         # Convert
         img = img.transpose((2, 0, 1))[::-1]  # HWC to CHW, BGR to RGB
